requirements.txt
fastapi
uvicorn
scikit-learn
pandas
numpy
streamlit
requests
python-dotenv
requests
